requirements.txt
# 🤖 Oracle Trading Bot - Dependencies
# pip install -r requirements.txt --break-system-packages

# === Core ===
python-telegram-bot>=20.0
aiohttp>=3.9.0

# === Web Server ===
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0  # Production server

# === Google Sheets ===
gspread>=6.0.0
google-auth>=2.25.0

# === Utils ===
requests>=2.31.0
python-dotenv>=1.0.0

# === Optional: Price APIs ===
# ccxt>=4.0.0           # Pour prix crypto
# yfinance>=0.2.0       # Pour prix stocks/forex

# === Development ===
# pytest>=7.0.0
# black>=23.0.0
